feat: JRD idempotent job setting & granular exception logging(RDCC-1404)
