+++ build.gradle
@@ -1,9 +1,14 @@
 buildscript {
+    ext {
+        springBootVersion = '2.2.6.RELEASE'
+    }
     repositories {
         jcenter()
+        mavenCentral()
     }
     dependencies {
         classpath("net.serenity-bdd:serenity-gradle-plugin:2.0.11")
+        classpath("org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}")
     }
 }
 
@@ -17,7 +22,6 @@ plugins {
     id 'io.spring.dependency-management' version '1.0.9.RELEASE'
     id 'org.owasp.dependencycheck' version '5.3.0'
     id 'org.sonarqube' version '2.8'
-    id 'org.springframework.boot' version '2.2.6.RELEASE'
     id "org.flywaydb.flyway" version "6.2.0"
 }
 
@@ -25,6 +29,7 @@ apply plugin: 'java'
 apply plugin: 'net.serenity-bdd.aggregator'
 apply plugin: 'project-report'
 apply plugin: 'idea'
+apply plugin: 'org.springframework.boot'
 
 def versions = [
         lombok             : '1.18.6',
@@ -35,10 +40,10 @@ def versions = [
         reformS2sClient    : '3.0.0',
         serenity           : '2.0.23',
         sonarPitest        : '0.5',
-        springBoot         : '2.1.4.RELEASE',
+        springBoot         : '2.2.6.RELEASE',
         springHystrix      : '2.1.1.RELEASE',
         springfoxSwagger   : '2.9.2',
-        camel              : '2.25.0'
+        camel              : '2.25.1'
 ]
 
 mainClassName = 'uk.gov.hmcts.reform.juddata.JudicialApplication'
@@ -154,7 +159,7 @@ task migratePostgresDatabase(type: org.flywaydb.gradle.task.FlywayMigrateTask) {
 }
 
 jacoco {
-    toolVersion = "0.8.1"
+    toolVersion = "0.8.5"
 }
 
 jacocoTestReport {
@@ -273,9 +278,10 @@ dependencies {
     compile group: 'org.apache.camel', name: 'camel-spring-boot-starter', version: versions.camel
 
     compile group: 'org.apache.camel', name: 'camel-core', version: versions.camel
-    compile group: 'org.apache.camel', name: 'spi-annotations', version: '3.0.0'
+    compile group: 'org.apache.camel', name: 'spi-annotations', version: '3.3.0'
 
     compile group: 'com.opencsv', name: 'opencsv', version: '3.7'
+    compile group: 'commons-lang', name: 'commons-lang', version: '2.6'
 
     compileOnly group: 'org.projectlombok', name: 'lombok', version: versions.lombok
     annotationProcessor group: 'org.projectlombok', name: 'lombok', version: versions.lombok
