--- src/integrationTest/java/uk/gov/hmcts/reform/juddata/cameltest/LeafRouteAbstractTest.java
@@ -1,42 +1,0 @@
-package uk.gov.hmcts.reform.juddata.cameltest;
-
-import org.apache.camel.CamelContext;
-import org.apache.camel.ProducerTemplate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.jdbc.core.JdbcTemplate;
-import uk.gov.hmcts.reform.juddata.camel.route.LeafTableRoute;
-import uk.gov.hmcts.reform.juddata.camel.util.DataLoadUtil;
-
-public abstract class LeafRouteAbstractTest {
-
-    @Autowired
-    protected CamelContext camelContext;
-
-    @Autowired
-    protected JdbcTemplate jdbcTemplate;
-
-    @Autowired
-    protected LeafTableRoute leafTableRoute;
-
-    @Value("${base-location-select-jrd-sql}")
-    protected String baseLocationSql;
-
-    @Value("${region-select-jrd-sql}")
-    protected String regionSql;
-
-    @Value("${contract-select-jrd-sql}")
-    protected String contractSql;
-
-    @Value("${role-select-jrd-sql}")
-    protected String roleSql;
-
-    @Autowired
-    protected ProducerTemplate producerTemplate;
-
-    @Value("${start-leaf-route}")
-    protected String startLeafRoute;
-
-    @Autowired
-    protected DataLoadUtil dataLoadUtil;
-}
