--- src/integrationTest/java/uk/gov/hmcts/reform/juddata/cameltest/ParentOrchestrationRouteValidationTest.java
@@ -1,125 +1,0 @@
-package uk.gov.hmcts.reform.juddata.cameltest;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static uk.gov.hmcts.reform.juddata.camel.util.MappingConstants.JUDICIAL_USER_PROFILE_ORCHESTRATION;
-import static uk.gov.hmcts.reform.juddata.camel.util.MappingConstants.ORCHESTRATED_ROUTE;
-import static uk.gov.hmcts.reform.juddata.cameltest.testsupport.IntegrationTestSupport.setSourcePath;
-import static uk.gov.hmcts.reform.juddata.cameltest.testsupport.ParentIntegrationTestSupport.fileWithInvalidHeader;
-import static uk.gov.hmcts.reform.juddata.cameltest.testsupport.ParentIntegrationTestSupport.fileWithInvalidJsr;
-import static uk.gov.hmcts.reform.juddata.cameltest.testsupport.ParentIntegrationTestSupport.fileWithInvalidJsrExceedsThreshold;
-import static uk.gov.hmcts.reform.juddata.cameltest.testsupport.ParentIntegrationTestSupport.setSourceData;
-
-import java.util.List;
-import java.util.Map;
-
-import org.apache.camel.test.spring.CamelSpringRunner;
-import org.apache.camel.test.spring.CamelTestContextBootstrapper;
-import org.apache.camel.test.spring.MockEndpoints;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.data.jpa.JpaRepositoriesAutoConfiguration;
-import org.springframework.boot.test.context.ConfigFileApplicationContextInitializer;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.context.jdbc.SqlConfig;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-import uk.gov.hmcts.reform.juddata.config.ParentCamelConfig;
-
-@TestPropertySource(properties = {"spring.config.location=classpath:application-integration.yml"})
-@RunWith(CamelSpringRunner.class)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-@MockEndpoints("log:*")
-@ContextConfiguration(classes = {ParentCamelConfig.class, CamelTestContextBootstrapper.class}, initializers = ConfigFileApplicationContextInitializer.class)
-@SpringBootTest
-@EnableAutoConfiguration(exclude = JpaRepositoriesAutoConfiguration.class)
-@EnableTransactionManagement
-@SqlConfig(dataSource =  "dataSource", transactionManager = "txManager", transactionMode = SqlConfig.TransactionMode.ISOLATED)
-public class ParentOrchestrationRouteValidationTest extends ParentRouteAbstractTest {
-
-    @Value("${exception-select-query}")
-    String exceptionQuery;
-
-
-    @BeforeClass
-    public static void beforeAll() throws Exception {
-        setSourcePath("classpath:archivalFiles", "archival.path");
-        setSourcePath("classpath:sourceFiles", "active.path");
-    }
-
-    @Before
-    public void init() {
-        camelContext.getGlobalOptions().put(ORCHESTRATED_ROUTE, JUDICIAL_USER_PROFILE_ORCHESTRATION);
-        dataLoadUtil.setGlobalConstant(camelContext, JUDICIAL_USER_PROFILE_ORCHESTRATION);
-    }
-
-    @Test
-    @Sql(scripts = {"/testData/truncate-parent.sql","/testData/truncate-exception.sql","/testData/default-leaf-load.sql"})
-    public void testParentOrchestrationInvalidHeaderRollback() throws Exception {
-        setSourceData(fileWithInvalidHeader);
-        parentRoute.startRoute();
-        producerTemplate.sendBody(startRoute, "test JRD orchestration");
-
-        List<Map<String, Object>> judicialUserProfileList = jdbcTemplate.queryForList(sql);
-        assertEquals(judicialUserProfileList.size(), 0);
-
-        List<Map<String, Object>> judicialAppointmentList = jdbcTemplate.queryForList(sqlChild1);
-        assertEquals(judicialAppointmentList.size(), 0);
-
-        List<Map<String, Object>> exceptionList = jdbcTemplate.queryForList(exceptionQuery);
-        assertNotNull(exceptionList.get(0).get("file_name"));
-        assertNotNull(exceptionList.get(0).get("scheduler_start_time"));
-        assertNotNull(exceptionList.get(0).get("error_description"));
-        assertNotNull(exceptionList.get(0).get("updated_timestamp"));
-        assertEquals(exceptionList.size(), 1);
-    }
-
-    @Test
-    @Sql(scripts = {"/testData/truncate-parent.sql","/testData/truncate-exception.sql","/testData/default-leaf-load.sql"})
-    public void testParentOrchestrationJsrAuditTest() throws Exception {
-        setSourceData(fileWithInvalidJsr);
-        parentRoute.startRoute();
-        producerTemplate.sendBody(startRoute, "test JRD orchestration");
-
-        List<Map<String, Object>> judicialUserProfileList = jdbcTemplate.queryForList(sql);
-        assertEquals(judicialUserProfileList.get(0).get("elinks_id"), "1");
-        assertEquals(judicialUserProfileList.get(1).get("elinks_id"), "2");
-        assertEquals(judicialUserProfileList.get(0).get("email_id"), "joe.bloggs@ejudiciary.net");
-        assertEquals(judicialUserProfileList.get(1).get("email_id"), "jo1e.bloggs@ejudiciary.net");
-        assertEquals(judicialUserProfileList.size(), 2);
-
-        List<Map<String, Object>> judicialAppointmentList = jdbcTemplate.queryForList(sqlChild1);
-        assertNotNull(judicialAppointmentList.get(0).get("judicial_office_appointment_id"));
-        assertNotNull(judicialAppointmentList.get(0).get("judicial_office_appointment_id"));
-        assertEquals(judicialAppointmentList.get(0).get("elinks_id"), "1");
-        assertEquals(judicialAppointmentList.get(1).get("elinks_id"), "2");
-        assertEquals(judicialAppointmentList.size(), 2);
-
-        List<Map<String, Object>> exceptionList = jdbcTemplate.queryForList(exceptionQuery);
-
-        for (int count = 0; count < 6; count++) {
-            assertNotNull(exceptionList.get(0).get("table_name"));
-            assertNotNull(exceptionList.get(0).get("scheduler_start_time"));
-            assertNotNull(exceptionList.get(0).get("key"));
-            assertNotNull(exceptionList.get(0).get("field_in_error"));
-            assertNotNull(exceptionList.get(0).get("error_description"));
-            assertNotNull(exceptionList.get(0).get("updated_timestamp"));
-        }
-        assertEquals(exceptionList.size(), 6);
-    }
-
-    @Test(expected = Exception.class)
-    @Sql(scripts = {"/testData/truncate-parent.sql", "/testData/truncate-exception.sql","/testData/default-leaf-load.sql"})
-    public void testParentOrchestrationJsrExceedsThresholdAuditTest() throws Exception {
-        setSourceData(fileWithInvalidJsrExceedsThreshold);
-        parentRoute.startRoute();
-        producerTemplate.sendBody(startRoute, "test JRD orchestration");
-    }
-}
