+++ src/integrationTest/java/uk/gov/hmcts/reform/juddata/config/ParentCamelConfig.java
@@ -30,6 +30,7 @@ import uk.gov.hmcts.reform.juddata.camel.processor.JudicialUserProfileProcessor;
 import uk.gov.hmcts.reform.juddata.camel.route.ParentOrchestrationRoute;
 import uk.gov.hmcts.reform.juddata.camel.service.AuditProcessingService;
 import uk.gov.hmcts.reform.juddata.camel.service.EmailService;
+import uk.gov.hmcts.reform.juddata.camel.task.IdempotentTask;
 import uk.gov.hmcts.reform.juddata.camel.task.LeafRouteTask;
 import uk.gov.hmcts.reform.juddata.camel.task.ParentRouteTask;
 import uk.gov.hmcts.reform.juddata.camel.util.DataLoadUtil;
@@ -202,6 +203,11 @@ public class ParentCamelConfig {
     @Bean
     LeafRouteTask leafRouteTask() {
         return new LeafRouteTask();
+    }
+
+    @Bean
+    IdempotentTask idempotentTask() {
+        return new IdempotentTask();
     }
 
 }
