+++ src/integrationTest/resources/testData/truncate-audit.sql
@@ -1,0 +1,2 @@
+truncate dataload_schedular_audit;
+commit;
