+++ src/main/java/uk/gov/hmcts/reform/juddata/camel/processor/ExceptionProcessor.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.juddata.camel.processor;
 import static java.lang.Boolean.TRUE;
 import static java.util.Objects.isNull;
 import static org.apache.camel.Exchange.EXCEPTION_CAUGHT;
+import static uk.gov.hmcts.reform.juddata.camel.util.MappingConstants.ERROR_MESSAGE;
 import static uk.gov.hmcts.reform.juddata.camel.util.MappingConstants.FAILURE;
 import static uk.gov.hmcts.reform.juddata.camel.util.MappingConstants.IS_EXCEPTION_HANDLED;
 import static uk.gov.hmcts.reform.juddata.camel.util.MappingConstants.SCHEDULER_STATUS;
@@ -13,6 +14,7 @@ import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import uk.gov.hmcts.reform.juddata.camel.service.AuditProcessingService;
 import uk.gov.hmcts.reform.juddata.camel.service.EmailService;
 
 @Component
@@ -25,6 +27,9 @@ public class ExceptionProcessor implements Processor {
     @Autowired
     EmailService emailService;
 
+    @Autowired
+    AuditProcessingService auditProcessingService;
+
     @Override
     public void process(Exchange exchange) throws Exception {
 
@@ -33,8 +38,8 @@ public class ExceptionProcessor implements Processor {
             log.error("::::exception in route for data processing::::" + exception);
             exchange.getContext().getGlobalOptions().put(SCHEDULER_STATUS, FAILURE);
             exchange.getContext().getGlobalOptions().put(IS_EXCEPTION_HANDLED, TRUE.toString());
-            //check mail flag and send mail
-            emailService.sendEmail(exception.getMessage());
+            exchange.getContext().getGlobalOptions().put(ERROR_MESSAGE, exception.getMessage());
+            throw exception;
         }
     }
 }
