+++ src/main/java/uk/gov/hmcts/reform/juddata/camel/task/IdempotentTask.java
@@ -1,0 +1,19 @@
+package uk.gov.hmcts.reform.juddata.camel.task;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.batch.core.StepContribution;
+import org.springframework.batch.core.scope.context.ChunkContext;
+import org.springframework.batch.core.step.tasklet.Tasklet;
+import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.stereotype.Component;
+
+@Component
+@Slf4j
+public class IdempotentTask implements Tasklet {
+
+    @Override
+    public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
+        log.info("::No tasks will be run as it has been executed by previous job::");
+        return RepeatStatus.FINISHED;
+    }
+}
