+++ src/main/java/uk/gov/hmcts/reform/juddata/camel/util/MappingConstants.java
@@ -65,4 +65,6 @@ public class MappingConstants {
 
     public static final String SUCCESS = "Success";
 
+    public static final String ERROR_MESSAGE = "ErrorMessage";
+
 }
