+++ src/main/java/uk/gov/hmcts/reform/juddata/configuration/BatchConfig.java
@@ -7,7 +7,6 @@ import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
 import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
-import org.springframework.batch.core.job.SimpleJob;
 import org.springframework.batch.core.launch.support.RunIdIncrementer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -15,6 +14,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
 import uk.gov.hmcts.reform.juddata.camel.service.AuditProcessingService;
+import uk.gov.hmcts.reform.juddata.camel.task.IdempotentTask;
 import uk.gov.hmcts.reform.juddata.camel.task.LeafRouteTask;
 import uk.gov.hmcts.reform.juddata.camel.task.ParentRouteTask;
 
@@ -46,6 +46,9 @@ public class BatchConfig {
     @Autowired
     LeafRouteTask leafRouteTask;
 
+    @Autowired
+    IdempotentTask idempotentTask;
+
 
     @Bean
     public Step stepLeafRoute() {
@@ -62,6 +65,13 @@ public class BatchConfig {
     }
 
     @Bean
+    public Step stepIdempotent() {
+        return steps.get("idempotent")
+                .tasklet(idempotentTask)
+                .build();
+    }
+
+    @Bean
     @Lazy
     public Job runRoutesJob() {
 
@@ -73,6 +83,6 @@ public class BatchConfig {
                     .build();
         }
 
-        return new SimpleJob();
+        return jobs.get(jobName).start(stepIdempotent()).build();
     }
 }
